requirements.txt
requests
transformers
pymongo
python-dotenv
pytest